test: rename files split tests
