+++ app/flows/fluggastrechteVorabcheck/__test__/testcasesErfolg.ts
@@ -1,0 +1,127 @@
+import { createMachine } from "xstate";
+import type { TestCases } from "~/flows/__test__/TestCases";
+import type { FluggastrechtVorabcheckContext } from "~/flows/fluggastrechteVorabcheck/context";
+import fluggastrechte from "~/flows/fluggastrechteVorabcheck/flow.json";
+import { guards } from "~/flows/fluggastrechteVorabcheck/guards";
+import type { FlowStateMachine } from "~/services/flow/server/buildFlowController";
+
+const machine: FlowStateMachine = createMachine(
+  { ...fluggastrechte, context: {} },
+  { guards },
+);
+
+const cases = [
+  [
+    {
+      bereich: "verspaetet",
+      verspaetung: "yes",
+      gruende: "yes",
+      verjaehrung: "yes",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "AMS",
+      endAirport: "MUC",
+      fluggesellschaft: "LH",
+    },
+    [
+      "start",
+      "bereich",
+      "verspaetung",
+      "gruende",
+      "gruende-hinweis",
+      "verjaehrung",
+      "flughaefen",
+      "fluggesellschaft",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg",
+    ],
+  ],
+  [
+    {
+      bereich: "verspaetet",
+      verspaetung: "yes",
+      gruende: "yes",
+      verjaehrung: "yes",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "BER",
+      endAirport: "MUC",
+      fluggesellschaft: "LH",
+    },
+    [
+      "start",
+      "bereich",
+      "verspaetung",
+      "gruende",
+      "gruende-hinweis",
+      "verjaehrung",
+      "flughaefen",
+      "fluggesellschaft",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg",
+    ],
+  ],
+  [
+    {
+      bereich: "verspaetet",
+      verspaetung: "yes",
+      gruende: "yes",
+      verjaehrung: "yes",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "BER",
+      endAirport: "MUC",
+      fluggesellschaft: "DL",
+    },
+    [
+      "start",
+      "bereich",
+      "verspaetung",
+      "gruende",
+      "gruende-hinweis",
+      "verjaehrung",
+      "flughaefen",
+      "fluggesellschaft",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg",
+    ],
+  ],
+] as const satisfies TestCases<FluggastrechtVorabcheckContext>;
+
+export const testCasesFluggastrechteErfolg = {
+  machine,
+  cases,
+};
