+++ app/flows/fluggastrechteVorabcheck/__test__/testcasesErfolgAnalog.ts
@@ -17,10 +17,16 @@ const cases = [
       verspaetung: "yes",
       gruende: "yes",
       verjaehrung: "yes",
-      startAirport: "JFK",
-      endAirport: "BER",
-      fluggesellschaft: "AF",
-      checkin: "no",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "AMS",
+      endAirport: "DRS",
+      fluggesellschaft: "LH",
     },
     [
       "start",
@@ -32,7 +38,13 @@ const cases = [
       "flughaefen",
       "fluggesellschaft",
       "checkin",
-      "ergebnis/checkin-abbruch",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg-analog",
     ],
   ],
   [
@@ -40,10 +52,36 @@ const cases = [
       bereich: "verspaetet",
       verspaetung: "yes",
       gruende: "yes",
-      startAirport: "GRU",
-      endAirport: "JFK",
+      verjaehrung: "yes",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "DRS",
+      endAirport: "ERF",
+      fluggesellschaft: "LH",
     },
-    ["flughaefen", "ergebnis/flughaefen-abbruch"],
+    [
+      "start",
+      "bereich",
+      "verspaetung",
+      "gruende",
+      "gruende-hinweis",
+      "verjaehrung",
+      "flughaefen",
+      "fluggesellschaft",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg-analog",
+    ],
   ],
   [
     {
@@ -51,8 +89,15 @@ const cases = [
       verspaetung: "yes",
       gruende: "yes",
       verjaehrung: "yes",
-      startAirport: "JFK",
-      endAirport: "BER",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "DRS",
+      endAirport: "ERF",
       fluggesellschaft: "DL",
     },
     [
@@ -64,7 +109,14 @@ const cases = [
       "verjaehrung",
       "flughaefen",
       "fluggesellschaft",
-      "ergebnis/fluggesellschaft-nicht-eu-abbruch",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg-analog",
     ],
   ],
   [
@@ -73,30 +125,15 @@ const cases = [
       verspaetung: "yes",
       gruende: "yes",
       verjaehrung: "yes",
-      startAirport: "CDG",
-      endAirport: "DRS",
-      fluggesellschaft: "DL",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "ergebnis/fluggesellschaft-nicht-eu-abbruch",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      startAirport: "CDG",
-      endAirport: "DRS",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "DRS",
+      endAirport: "ERF",
       fluggesellschaft: "sonstiges",
     },
     [
@@ -108,7 +145,14 @@ const cases = [
       "verjaehrung",
       "flughaefen",
       "fluggesellschaft",
-      "ergebnis/fluggesellschaft-abbruch",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg-analog",
     ],
   ],
   [
@@ -117,8 +161,15 @@ const cases = [
       verspaetung: "yes",
       gruende: "yes",
       verjaehrung: "yes",
-      startAirport: "AMS",
-      endAirport: "BER",
+      checkin: "yes",
+      kostenlos: "no",
+      rabatt: "no",
+      buchung: "yes",
+      abtretung: "no",
+      entschaedigung: "yes",
+      gericht: "no",
+      startAirport: "BER",
+      endAirport: "MUC",
       fluggesellschaft: "sonstiges",
     },
     [
@@ -130,56 +181,19 @@ const cases = [
       "verjaehrung",
       "flughaefen",
       "fluggesellschaft",
-      "ergebnis/fluggesellschaft-abbruch",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      startAirport: "JFK",
-      endAirport: "AMS",
-      fluggesellschaft: "sonstiges",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "ergebnis/fluggesellschaft-abbruch-eu",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      startAirport: "CDG",
-      endAirport: "BER",
-      fluggesellschaft: "DL",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "ergebnis/fluggesellschaft-nicht-eu-abbruch",
+      "checkin",
+      "kostenlos",
+      "rabatt",
+      "buchung",
+      "abtretung",
+      "entschaedigung",
+      "gericht",
+      "ergebnis/erfolg-analog",
     ],
   ],
 ] as const satisfies TestCases<FluggastrechtVorabcheckContext>;
 
-export const testCasesFluggastrechteVerspaetet = {
+export const testcasesFluggastrechteErfolgAnalog = {
   machine,
   cases,
 };
