--- app/flows/fluggastrechteVorabcheck/__test__/testcasesSuccessful.ts
@@ -1,561 +1,0 @@
-import { createMachine } from "xstate";
-import type { TestCases } from "~/flows/__test__/TestCases";
-import type { FluggastrechtVorabcheckContext } from "~/flows/fluggastrechteVorabcheck/context";
-import fluggastrechte from "~/flows/fluggastrechteVorabcheck/flow.json";
-import { guards } from "~/flows/fluggastrechteVorabcheck/guards";
-import type { FlowStateMachine } from "~/services/flow/server/buildFlowController";
-
-const machine: FlowStateMachine = createMachine(
-  { ...fluggastrechte, context: {} },
-  { guards },
-);
-
-const cases = [
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      startAirport: "JFK",
-      endAirport: "BER",
-      fluggesellschaft: "AF",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "no",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "ergebnis/erfolg-kontakt",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      startAirport: "JFK",
-      endAirport: "BER",
-      fluggesellschaft: "AF",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "yes",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-gericht",
-    ],
-  ],
-  [
-    {
-      bereich: "annullierung",
-      ankuendigung: "between7And13Days",
-      ersatzflug: "yes",
-      ersatzflugStartenZweiStunden: "yes",
-      ersatzflugLandenVierStunden: "no",
-      vertretbareGruendeAnnullierung: "no",
-      verjaehrung: "yes",
-      startAirport: "BER",
-      endAirport: "MUN",
-      fluggesellschaft: "AF",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "yes",
-    },
-    [
-      "start",
-      "bereich",
-      "ankuendigung",
-      "ersatzflug",
-      "ersatzflug-starten-zwei-stunden",
-      "ersatzflug-landen-vier-stunden",
-      "vertretbare-gruende-annullierung",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-gericht",
-    ],
-  ],
-  [
-    {
-      bereich: "nichtbefoerderung",
-      ausgleich: "yes",
-      ausgleichAngenommen: "no",
-      vertretbareGruende: "no",
-      startAirport: "JFK",
-      endAirport: "BER",
-      fluggesellschaft: "AF",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "yes",
-    },
-    [
-      "start",
-      "bereich",
-      "ausgleich",
-      "ausgleich-angenommen",
-      "checkin-nicht-befoerderung",
-      "vertretbare-gruende",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-gericht",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "AMS",
-      endAirport: "CDG",
-      fluggesellschaft: "LH",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-eu",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "AMS",
-      endAirport: "CDG",
-      fluggesellschaft: "DL",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-eu",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "AMS",
-      endAirport: "CDG",
-      fluggesellschaft: "sonstiges",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-eu",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "AMS",
-      endAirport: "DRS",
-      fluggesellschaft: "LH",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-analog",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "DRS",
-      endAirport: "ERF",
-      fluggesellschaft: "LH",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-analog",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "DRS",
-      endAirport: "ERF",
-      fluggesellschaft: "DL",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-analog",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "DRS",
-      endAirport: "ERF",
-      fluggesellschaft: "sonstiges",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-analog",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "BER",
-      endAirport: "MUC",
-      fluggesellschaft: "sonstiges",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg-analog",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "AMS",
-      endAirport: "MUC",
-      fluggesellschaft: "LH",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "BER",
-      endAirport: "MUC",
-      fluggesellschaft: "LH",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg",
-    ],
-  ],
-  [
-    {
-      bereich: "verspaetet",
-      verspaetung: "yes",
-      gruende: "yes",
-      verjaehrung: "yes",
-      checkin: "yes",
-      kostenlos: "no",
-      rabatt: "no",
-      buchung: "yes",
-      abtretung: "no",
-      entschaedigung: "yes",
-      gericht: "no",
-      startAirport: "BER",
-      endAirport: "MUC",
-      fluggesellschaft: "DL",
-    },
-    [
-      "start",
-      "bereich",
-      "verspaetung",
-      "gruende",
-      "gruende-hinweis",
-      "verjaehrung",
-      "flughaefen",
-      "fluggesellschaft",
-      "checkin",
-      "kostenlos",
-      "rabatt",
-      "buchung",
-      "abtretung",
-      "entschaedigung",
-      "gericht",
-      "ergebnis/erfolg",
-    ],
-  ],
-] as const satisfies TestCases<FluggastrechtVorabcheckContext>;
-
-export const testCasesFluggastrechteSuccessful = {
-  machine,
-  cases,
-};
